+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/checkpoint/ProcessCheckpointResource.java
@@ -25,17 +25,15 @@ import com.walmartlabs.concord.common.TemporaryPath;
 import com.walmartlabs.concord.server.ConcordApplicationException;
 import com.walmartlabs.concord.server.MultipartUtils;
 import com.walmartlabs.concord.server.metrics.WithTimer;
-import com.walmartlabs.concord.server.org.ResourceAccessLevel;
-import com.walmartlabs.concord.server.org.project.ProjectAccessManager;
-import com.walmartlabs.concord.server.process.*;
-import com.walmartlabs.concord.server.process.queue.ProcessQueueDao;
+import com.walmartlabs.concord.server.process.ProcessEntry;
+import com.walmartlabs.concord.server.process.ProcessKey;
+import com.walmartlabs.concord.server.process.ProcessManager;
+import com.walmartlabs.concord.server.process.ResumeProcessResponse;
 import com.walmartlabs.concord.server.process.state.ProcessCheckpointManager;
-import com.walmartlabs.concord.server.security.UserPrincipal;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.Authorization;
-import org.apache.shiro.authz.UnauthorizedException;
 import org.jboss.resteasy.plugins.providers.multipart.MultipartInput;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -48,12 +46,12 @@ import javax.inject.Singleton;
 import javax.validation.Valid;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
-import java.util.*;
+import java.util.List;
+import java.util.UUID;
 
 
 @Named
@@ -64,26 +62,14 @@ public class ProcessCheckpointResource implements Resource {
 
     private static final Logger log = LoggerFactory.getLogger(ProcessCheckpointResource.class);
 
-    private static final Set<ProcessStatus> RESTORE_ALLOWED_STATUSES = new HashSet<>(Arrays.asList(
-            ProcessStatus.FAILED, ProcessStatus.FINISHED, ProcessStatus.SUSPENDED, ProcessStatus.TIMED_OUT));
-
     private final ProcessManager processManager;
-    private final PayloadManager payloadManager;
     private final ProcessCheckpointManager checkpointManager;
-    private final ProcessQueueDao processQueueDao;
-    private final ProjectAccessManager projectAccessManager;
 
     @Inject
     public ProcessCheckpointResource(ProcessManager processManager,
-                                     PayloadManager payloadManager,
-                                     ProcessCheckpointManager checkpointManager,
-                                     ProcessQueueDao processQueueDao,
-                                     ProjectAccessManager projectAccessManager) {
+                                     ProcessCheckpointManager checkpointManager) {
         this.processManager = processManager;
-        this.payloadManager = payloadManager;
         this.checkpointManager = checkpointManager;
-        this.processQueueDao = processQueueDao;
-        this.projectAccessManager = projectAccessManager;
     }
 
     @GET
@@ -92,10 +78,10 @@ public class ProcessCheckpointResource implements Resource {
     @Produces(MediaType.APPLICATION_JSON)
     @WithTimer
     public List<ProcessCheckpointEntry> list(@ApiParam @PathParam("id") UUID instanceId) {
-        ProcessEntry entry = assertProcess(instanceId);
+        ProcessEntry entry = processManager.assertProcess(instanceId);
         ProcessKey processKey = ProcessKey.from(entry);
 
-        assertProcessCheckpointAccess(entry);
+        checkpointManager.assertProcessAccess(entry);
 
         return checkpointManager.list(processKey);
     }
@@ -112,32 +98,11 @@ public class ProcessCheckpointResource implements Resource {
 
         UUID checkpointId = request.getId();
 
-        ProcessEntry entry = assertProcess(instanceId);
+        ProcessEntry entry = processManager.assertProcess(instanceId);
         ProcessKey processKey = ProcessKey.from(entry);
 
-        assertProcessCheckpointAccess(entry);
+        processManager.restoreFromCheckpoint(processKey, checkpointId);
 
-        ProcessStatus s = entry.status();
-        if (!RESTORE_ALLOWED_STATUSES.contains(s)) {
-            throw new ConcordApplicationException("Unable to restore a checkpoint, the process is " + s);
-        }
-
-        String eventName = checkpointManager.restoreCheckpoint(processKey, checkpointId);
-        if (eventName == null) {
-            throw new ConcordApplicationException("Checkpoint " + checkpointId + " not found");
-        }
-
-        Payload payload;
-        try {
-            payload = payloadManager.createResumePayload(processKey, eventName, null);
-        } catch (IOException e) {
-            log.error("restore ['{}', '{}'] -> error creating a payload: {}", processKey, eventName, e);
-            throw new ConcordApplicationException("Error creating a payload", e);
-        }
-
-        processQueueDao.updateStatus(processKey, ProcessStatus.SUSPENDED, Collections.singletonMap("checkpointId", checkpointId));
-
-        processManager.resume(payload);
         return new ResumeProcessResponse();
     }
 
@@ -147,7 +112,7 @@ public class ProcessCheckpointResource implements Resource {
     public void uploadCheckpoint(@PathParam("id") UUID instanceId,
                                  @ApiParam MultipartInput input) {
 
-        ProcessEntry entry = assertProcess(instanceId);
+        ProcessEntry entry = processManager.assertProcess(instanceId);
         ProcessKey processKey = ProcessKey.from(entry);
 
         UUID checkpointId = MultipartUtils.getUuid(input, "id");
@@ -163,36 +128,5 @@ public class ProcessCheckpointResource implements Resource {
         }
 
         log.info("uploadCheckpoint ['{}'] -> done", processKey);
-    }
-
-    private ProcessEntry assertProcess(UUID instanceId) {
-        ProcessEntry p = processQueueDao.get(PartialProcessKey.from(instanceId));
-        if (p == null) {
-            throw new ConcordApplicationException("Process instance not found", Response.Status.NOT_FOUND);
-        }
-        return p;
-    }
-
-    private void assertProcessCheckpointAccess(ProcessEntry p) {
-        UserPrincipal principal = UserPrincipal.assertCurrent();
-
-        UUID initiatorId = p.initiatorId();
-        if (principal.getId().equals(initiatorId)) {
-            // process owners should be able to restore the process from a checkpoint
-            return;
-        }
-
-        if (principal.isAdmin()) {
-            return;
-        }
-
-        UUID projectId = p.projectId();
-        if (projectId != null) {
-            projectAccessManager.assertProjectAccess(projectId, ResourceAccessLevel.WRITER, false);
-            return;
-        }
-
-        throw new UnauthorizedException("The current user (" + principal.getUsername() + ") doesn't have " +
-                "the necessary permissions to restore the process using a checkpoint: " + p.instanceId());
     }
 }